test: share browser layout style installer
